requirements.txt
flask
flask-cors
google-generativeai>=0.4.1
requests
gunicorn
gTTS
